addressing PR feedback
1. refactored login_required to remove code duplication
2. eliminated author lookup for each post, now single find() operation
3. eliminated wildcard/regex searching for users, now exact match only
