+++ usecases/social_media/.gitignore
@@ -19,3 +19,4 @@ Thumbs.db
 
 # Other
 CLAUDE.md
+.claude/
