+++ usecases/social_media/routes/users.py
@@ -1,18 +1,9 @@
 from flask import Blueprint, render_template, request, redirect, url_for, session, flash
 from models import user as user_model
 from models import post as post_model
+from utils.decorators import login_required
 
 bp = Blueprint('users', __name__)
-
-def login_required(f):
-    """Decorator to require login for a route."""
-    def decorated_function(*args, **kwargs):
-        if 'user_id' not in session:
-            flash('Please login to access this page', 'error')
-            return redirect(url_for('auth.login'))
-        return f(*args, **kwargs)
-    decorated_function.__name__ = f.__name__
-    return decorated_function
 
 @bp.route('/profile/<username>')
 @login_required
