+++ usecases/social_media/utils/decorators.py
@@ -1,0 +1,11 @@
+from flask import session, flash, redirect, url_for
+
+def login_required(f):
+    """Decorator to require login for a route."""
+    def decorated_function(*args, **kwargs):
+        if 'user_id' not in session:
+            flash('Please login to access this page', 'error')
+            return redirect(url_for('auth.login'))
+        return f(*args, **kwargs)
+    decorated_function.__name__ = f.__name__
+    return decorated_function
